feat: add product page skeleton and rearrange skeletons
