+++ src/app/pages/catalog-page/catalogPageSkeleton.tsx
