--- src/components/ui/page-skeleton/productPageSkeleton.tsx
@@ -1,13 +1,0 @@
-import { VISIBILITY_DELAY } from '~config/constants';
-import { useTimeout } from '~hooks/useTimeout';
-import { useState } from 'react';
-
-export const ProductPageSkeleton = () => {
-  const [isVisible, setIsVisible] = useState(false);
-
-  useTimeout(() => {
-    setIsVisible(true);
-  }, VISIBILITY_DELAY);
-
-  return isVisible && <></>;
-};
